102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  * Return: Always 0.
  */
 int main(void){
-int i = 1, j = 2, k, count = 50;
+long long i = 1, j = 2, k, count = 50;
 printf("%d, %d, ", i, j);
 for (k = 0; k < count; k++)
 {
